Added a lot more Unit tests. Started work on the airplane WS

+++ QuietQuillBE/Application/Journals/Commands/DeleteEntry/DeleteEntryHandler.cs
@@ -11,8 +11,6 @@ public class DeleteEntryHandler : IRequestHandler<DeleteEntryCommand, bool>
     private readonly IJournalEntryRepository _journalEntryRepository;
     private readonly IValidator<DeleteEntryCommand> _validator;
     
-    
-    
     public DeleteEntryHandler(IUnitOfWork unitOfWork, IJournalEntryRepository journalEntryRepository, IValidator<DeleteEntryCommand> validator)
     {
         _unitOfWork = unitOfWork;
@@ -20,19 +18,19 @@ public class DeleteEntryHandler : IRequestHandler<DeleteEntryCommand, bool>
         _validator = validator;
     }
     
-    
-    
     public async Task<bool> Handle(DeleteEntryCommand request, CancellationToken cancellationToken)
     {
-        
-        var journalEntry =  _journalEntryRepository.DeleteJournalEntry(request.journalId);
-        if (journalEntry)
+        var validationResult = await _validator.ValidateAsync(request, cancellationToken);
+        if (!validationResult.IsValid)
         {
-            await _unitOfWork.SaveChangesAsync(cancellationToken);
-            return journalEntry;            
+            throw new ValidationException(validationResult.Errors);
         }
-        
-        
-        return journalEntry;
+        var journalEntryDeleted = _journalEntryRepository.DeleteJournalEntry(request.journalId);
+        if (!journalEntryDeleted)
+        {
+            return false;
+        }
+        await _unitOfWork.SaveChangesAsync(cancellationToken);
+        return true;
     }
 }
